use floats instead of doubles in nn
The original nn uses floats, so there is no point in wastin ~10MB for doubles. This change does not affect classification results.


Former-commit-id: 8ae6365678541e683876f24a273c38fd2f93f72a

+++ languagetool-core/src/main/java/org/languagetool/rules/neuralnetwork/Classifier.java
@@ -17,7 +17,7 @@ public class Classifier implements IClassifier {
         b_fc1 = new Matrix(bPath).transpose();
     }
 
-    public double[] getScores(String[] context) {
+    public float[] getScores(String[] context) {
         final Matrix x = new Matrix(Arrays.stream(context)
                 .map(dictionary::safeGet)
                 .map(embedding::row)
@@ -26,10 +26,10 @@ public class Classifier implements IClassifier {
         return x.mul(W_fc1).add(b_fc1).row(0);
     }
 
-    private static double[] concat(double[] a, double[] b) {
+    private static float[] concat(float[] a, float[] b) {
         int aLen = a.length;
         int bLen = b.length;
-        double[] c= new double[aLen + bLen];
+        float[] c= new float[aLen + bLen];
         System.arraycopy(a, 0, c, 0, aLen);
         System.arraycopy(b, 0, c, aLen, bLen);
         return c;
